requirements.txt
flask
flask-cors
google-generativeai
anthropic
openai==0.28
llm-client
gunicorn